Added source output & R9 detection
