+++ firmware/application/CMakeLists.txt
@@ -122,7 +122,6 @@ set(CSRC
 	${HACKRF_PATH}/firmware/hackrf_usb/usb_device.c
 	${HACKRF_PATH}/firmware/common/usb_request.c
 	${HACKRF_PATH}/firmware/common/usb_standard_request.c
-	shell.c
 	${CHIBIOS}/os/various/chprintf.c
 )
 
@@ -130,6 +129,7 @@ set(CSRC
 # setting.
 set(CPPSRC
 	main.cpp
+	shell.cpp
 	${COMMON}/acars_packet.cpp
 	${COMMON}/adsb.cpp
 	${COMMON}/adsb_frame.cpp
